# Faster in-memory weight transfer for transformers-neuronx