## Overview ##

This PR improves the latency of transferring model weights from the HF representation to the INF2 representation by doing it in-memory as opposed to using the disk as a buffer.  For OPT-66B on inf2.48xl, this improves the latency of weight transfer from 840s to 5s. 

## Description ## 

Splitting the model into small pieces on the disk and then loading it, reduces CPU memory requirements. However, this operation demands high IOPS from the disk. Typical EBS volumes used by EC2 (gp3) and SageMaker (gp2) do not provide high IOPS by default. This leads to EBS throttling, which manifests as a massive latency to transfer model weights (upto 1400 secs).

However, Large Language Models are typically used with large instances with large CPU memory. As a result, for instances like inf2.24xl, inf2.48xl, trn1.32xl, loading from disk is wasteful in terms of instance hours. Direct in-memory transfer is much faster (<5s) in these cases.

This PR modifies the load_model code path to first try in-memory loading and then fall back to disk loading in the event of an overflow.

## Testing ##
Tested in EC2 on inf2.48xl and inf2.24xl with OPT-66B and OPT-13B