@Lokiiiiii 2023-10-10
Additionally tested `facebook/opt-1.3b` and `OpenAssistant/llama2-13b-orca-8k-3319` on inf2.24xl (EC2)